feat: add endpoint to create or update post

+++ posts/views.py
@@ -1,7 +1,6 @@
-from django.shortcuts import render
-from rest_framework import status, generics
+from rest_framework import status, generics, permissions
 from posts.models import Post
-from .serializers import PostSerializer, CommentSerializer, PostAttachmentSerializer
+from .serializers import PostSerializer, CommentSerializer, PostAttachmentSerializer,PostCreateSerializer
 from rest_framework.views import APIView
 from core.responses import CustomResponse
 
@@ -17,13 +16,6 @@ class PostView(APIView):
             message="Posts retrieved successfully",
         )
 
-    def post(self, request):
-        serializer = PostSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return CustomResponse(data=serializer.data, status=status.HTTP_201_CREATED ,message="Post created successfully")
-        return CustomResponse(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST, message="Failed to create post")
-
 class PostDetailView(APIView):
     def get(self, request, pk):
         try:
@@ -33,21 +25,53 @@ class PostDetailView(APIView):
         except Post.DoesNotExist:
             return CustomResponse(status=status.HTTP_404_NOT_FOUND, message="Post not found")
 
-    def put(self, request, pk):
-        try:
-            post = Post.objects.get(pk=pk)
-            serializer = PostSerializer(post, data=request.data)
-            if serializer.is_valid():
-                serializer.save()
-                return CustomResponse(data=serializer.data, status=status.HTTP_200_OK ,message="Post updated successfully")
-            return CustomResponse(data=serializer.errors, status=400, message="Failed to update post")
-        except Post.DoesNotExist:
-            return CustomResponse(status=status.HTTP_404_NOT_FOUND, message="Post not found")
-
     def delete(self, request, pk):
+        if not request.user.is_authenticated:
+            return CustomResponse(
+                data=None,
+                status=status.HTTP_401_UNAUTHORIZED,
+                message="Authentication required to create a post",
+            )
         try:
             post = Post.objects.get(pk=pk)
             post.delete()
             return CustomResponse(data={} , status=status.HTTP_200_OK ,message="Post deleted successfully")
         except Post.DoesNotExist:
             return CustomResponse(status=status.HTTP_404_NOT_FOUND, message="Post not found")
+
+
+class PostCreateAPIView(generics.CreateAPIView):
+    queryset = Post.objects.all()
+    serializer_class = PostCreateSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+    def perform_create(self, serializer):
+        serializer.save(created_by=self.request.user)
+
+    def create(self, request, *args, **kwargs):
+        super().create(request, *args, **kwargs)  # Just create, no need to capture response
+        queryset = self.get_queryset().order_by("-created_at")  # or whatever ordering you want
+        serializer = PostSerializer(queryset, many=True, context={"request": request})
+        return CustomResponse(
+            data=serializer.data,
+            status=status.HTTP_201_CREATED,
+            message="Post created successfully",
+        )
+
+
+class PostUpdateAPIView(generics.UpdateAPIView):
+    queryset = Post.objects.all()
+    serializer_class = PostCreateSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+    def update(self, request, *args, **kwargs):
+        response = super().update(request, *args, **kwargs)
+        instance = (
+            self.get_object()
+        )  # Because in update, self.get_object() gives you the updated instance
+        serializer = PostSerializer(instance, context={"request": request})
+        return CustomResponse(
+            data=serializer.data,
+            status=status.HTTP_200_OK,
+            message="Post updated successfully",
+        )
